Add a real Case Studies page; the landing button pointed at Notices
Landing's "Case Studies" button linked to /notices — the same page as
the "Notices" button beside it — so there was no way to actually see the
verified public-source acquisition dataset already sitting unused in
data/real_acquisition_seed and app.real_seed. This wires it up instead of
just fixing the link.

Getting the seed to load surfaced three real bugs it had never hit before:
- app.real_seed resolves its CSV path two parents above itself, which is
  the monorepo root outside Docker but nothing inside the api container —
  docker-compose.yml now mounts ../data at /data to match.
- Two rows in projects.csv (BBC8, BBC-COMP) carry one stray extra comma
  each, shifting every field after it by one position — status ended up
  in area_ha as the literal string "compensation_reported". Fixed by
  removing the extra field so both rows line up with the header again.
- compensation_awarded/paid were Integer columns. The Upper Krishna
  Project row alone reports 5,440,000,000 (₹544 crore) — already past a
  32-bit column's ~2.1 billion ceiling by itself. Migration 0006 widens
  both to bigint.

/public-acquisitions and /public-acquisitions/summary drop their auth
requirement — everything in the table already traces to a gazette or a
published news report, the same reasoning /notices already stands on, and
a Case Studies page reached from the landing page has to work for a
visitor who hasn't signed in.

New CaseStudies.jsx renders each verified project (and the one
compensation milestone) as a card: department, location, notification,
area, compensation where publicly known, a same-project parcel count, and
the source citation — leaving a figure blank rather than estimating it
where the public record didn't provide one, per that data directory's own
integrity rules. A district filter and a one-line summary sit above the
list; parcel-level rows stay supporting detail rather than 21 more cards.

+++ app/models/public_records.py
@@ -1,6 +1,6 @@
 """Database model for curated public-source acquisition records."""
 
-from sqlalchemy import Boolean, Float, Integer, String, Text
+from sqlalchemy import BigInteger, Boolean, Float, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.database import Base
@@ -37,8 +37,11 @@ class PublicAcquisitionRecord(Base):
     area_acres: Mapped[float | None] = mapped_column(Float, nullable=True)
     owner_name_public: Mapped[str | None] = mapped_column(Text, nullable=True)
     owner_data_status: Mapped[str | None] = mapped_column(String(160), nullable=True)
-    compensation_awarded: Mapped[int | None] = mapped_column(Integer, nullable=True)
-    compensation_paid: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    # BigInteger, not Integer: a project-level figure like the Upper
+    # Krishna Project's reported ₹544 crore disbursed is 5,440,000,000 —
+    # already past a 32-bit column's ~2.1 billion ceiling on its own.
+    compensation_awarded: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
+    compensation_paid: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     payment_status: Mapped[str | None] = mapped_column(String(80), nullable=True)
     source: Mapped[str | None] = mapped_column(String(200), nullable=True)
     source_reference: Mapped[str | None] = mapped_column(Text, nullable=True)
